refactor!: relocate pure-data types from Editor/Services to Editor/Models
`Services/` is documented in `Beutl.Editor/CLAUDE.md` as the editing
pipeline directory, but it has accumulated payload types that have
nothing to do with the pipeline: `AudioFrameSnapshot` (audio-preview
PCM record), `CacheBlock` (frame-cache display block),
`ElementDescription` (placement parameter record). All three are
referenced from `Beutl.Editor.Components` views and the app
ViewModels, not from any service implementation.

Introduce `Beutl.Editor/Models/` and move the three records there with
namespace `Beutl.Editor.Models`. Service interfaces that previously
exposed them (`IBufferStatus.CacheBlocks`, `IPreviewPlayer.PlayAudio`,
`IElementAdder.AddElement`, `IElementClipboardService` paste-files
path via `ElementAdderImpl`) take a `using Beutl.Editor.Models;`
directive; test fixtures for the three types follow.

The split makes the `Services/` directory contract honest again — what
remains is interfaces, implementations, and the stateless helpers /
constants those services dispatch on (`DuplicateHelper`,
`ObjectRegenerator`, `BeutlClipboardFormats`, `IClipboardGateway` +
its `ClipboardEntry` record). The latter group is the working state
of the editing pipeline, not generic editor models, so it stays put.

BREAKING CHANGE: `Beutl.Editor.Services.AudioFrameSnapshot`,
`Beutl.Editor.Services.CacheBlock`, and
`Beutl.Editor.Services.ElementDescription` are renamed to
`Beutl.Editor.Models.{AudioFrameSnapshot|CacheBlock|ElementDescription}`.
Out-of-tree consumers add a `using Beutl.Editor.Models;` directive
(or fully qualify), nothing else changes.
