+++ src/Beutl.Editor.Components/AudioVisualizerTab/ViewModels/AudioVisualizerTabViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text.Json.Nodes;
+using Beutl.Editor.Models;
 using Beutl.Editor.Services;
 using Microsoft.Extensions.DependencyInjection;
 using Reactive.Bindings;
