+++ src/Beutl/Services/Tutorials/AnimationEditTutorial.cs
@@ -7,6 +7,7 @@ using Beutl.Editor.Components.GraphEditorTab;
 using Beutl.Editor.Components.Helpers;
 using Beutl.Editor.Components.LibraryTab;
 using Beutl.Editor.Components.LibraryTab.Views;
+using Beutl.Editor.Models;
 using Beutl.Engine;
 using Beutl.Graphics;
 using Beutl.Graphics.Shapes;
